oplog: Flush operations with timestamps less than or equal to max_tsc
wait_synchronize(), flush_finish_max_timestamp() and ops_before_max_tsc() are
all currently implemented such that they return the list of operations with
timestamps less than the specified 'max_tsc' value. Make that "less than or
equal to" instead, because we'll need those semantics shortly, as explained
below.

In future, when processing the metadata log, we will need to track the 2 parts
of 'rename' operations: rename_unlink and rename_link, which will both have
exactly the same timestamp. To link them together, we will need a way to get a
list of operations with timestamps less than or equal to a specified timestamp,
and hence this patch will come in handy. Also, this change doesn't hurt the
existing behavior of fsync.