scalefs: Remove obsolete dependency handling code in preparation for a revamp
In the earlier design, there were quite a few dependencies to be honored while
flushing out changes during an fsync() call. However, in the new design, we
have very few such dependencies to take care of (primarily related to rename)
and hence we won't need much of the infrastructure that was added to deal with
dependencies in the earlier design.

So remove those obsolete bits, to make way for the new design.