scalefs: Define a helper to apply an MFS operation and add it to the journal
Create a helper function to apply an MFS operation and add it to the in-memory
journal. Include the option of passing a transaction pointer. This will help
in cases where we wish to squeeze multiple different MFS operations together
into a single atomic transaction and apply them on the disk as a unit.
(Eg: The two parts [link and unlink] of a rename operation).

This helps prepare the ground for upcoming changes.