oplog: Convert sync_lock_ to a sleeplock to allow nesting a sleeplock inside it
The oplog's sync_lock_ is a spinlock at the moment. However, going ahead, we
will encounter a case where we acquire the oplog's sync_lock_ (via a call to
wait_synchronize()) and then try to acquire a sleeplock, with the sync_lock_
still held.

But a sleeplock cannot nest inside a spinlock (the scheduler panics if they
nest that way; this is the well-known "sleeping inside atomic section"
problem as per the Linux kernel terminology). So let us convert the oplog's
sync_lock_ into a sleeplock (as nesting a sleeplock inside another sleeplock
is legal).

This change doesn't seem to have any adverse effects on oplog's internal
health, at least in the testing so far!