scalefs: Add the initial rename dependency handling bits by identifying rename ops
The first step towards handling dependencies related to rename operations, is
to be able to identify the two parts (or sub-operations) that make up each
higher-level rename operation. (The two sub-operations being rename-link and
rename-unlink).

To achieve this, define an external stack called 'rename_stack', and teach
process_ops_from_oplog() to identify all the rename sub-operations that it
encounters and push their info onto the rename stack.

Also, add a new function called find_rename_op_counterpart() to add the other
directory (related to the rename operation) as a dependency for fsync, using
the pending stack. Once that mnode is added to the pending stack,
process_metadata_log() will naturally plow through its operations until it
finds the couterpart sub-operation of rename. Note that we don't have to fully
fsync the other directory, but only process it until we get to the other part
of the rename. Hence we choose the tsc value for the pending stack to be the
same as that of the rename operation.

By the way, the choice to build this infrastructure using external stacks
(pending_stack and rename_stack) rather than in-built function-call recursion
is very much intentional and carefully thought-out, as will be soon apparent
in subsequent commits (it helps avoid running out of kernel stack space due
to long chains of renames).