scalefs: Gather and apply the rename sub-operations (link/unlink) atomically
Since we keep track of all the rename sub-operations on a dedicated stack
(rename_stack), it follows that, upon working through all the dependent MFS
operations diligently, we will find the counterparts of each rename operation
next to each other on the rename_stack.

Once we detect this situation (two rename sub-operations making up a pair),
invoke apply_rename_pair() to apply both of them as a single indivisible
unit, using a common transaction. This is particularly important for directory
renames, because directories shouldn't be allowed to have more than 1 hard
link (and hence we can't perform the link and the unlink parts of the rename
separately in two different transactions).

We need to hold the appropriate locks while applying these operations. This
can get tricky here because these two operations belong to two different
directories (and hence two different oplogs). So institute a suitable lock
ordering rule to prevent deadlocks when running concurrent fsync()s on the
source and destination directories.

This commit completes the rename dependency handling under the new design.
Now that we have got most of the low-level details out of the way, here
are some highlights about the big-picture that are worth appreciating:

- An interesting use-case to keep in mind is this:

  mdir_1               mdir_2               mdir_3
  ------               ------               ------

  create               create               create
  link                 rename_unlink(t=1)   rename_link(t=1)
  unlink               link                 link
  rename_unlink(t=2)   rename_link(t=2)     link
  link                 link
  unlink
  link

  In this scenario, upon fsync()ing mdir_1, the correct minimal set of
  operations to apply, and their order is:

  [mdir_1 : create, link, unlink],
  [mdir_2 : create],
  [mdir_3 : create],
  [mdir_3 : rename_link(t=1) and mdir_2 : rename_unlink(t=1) together],
  [mdir_2 : link],
  [mdir_2 : rename_link(t=2) and mdir_1 : rename_unlink(t=2) together],
  [mdir_1 : link, unlink, link]

  Our new design honors this ordering and the minimalistic write-out effectively.

- We process both parts of the rename in a symmetric fashion, and apply them
  in a single atomic transaction; hence we will never encounter unbreakable
  dependency loops. Thus, we can have long rename chains, but irrespective of
  how convoluted they are, we will always find a unique execution order for
  all the pairs of rename operations.

- Since userspace can set up an unbounded chain of dependencies between
  rename operations (as shown above) all across the filesystem, it would be
  disastrous to use function-call recursion to handle dependencies, because
  we could easily run out of kernel stack space. That's why we use a carefully
  designed iterative algorithm, with external stacks to guide our path.

- The rename_stack stores the metadata extracted from the rename operations,
  rather than pointers to the operations themselves. This is very crucial,
  because, otherwise, a concurrent fsync on the other directory could apply
  and delete those operations, rendering the pointers stored on the
  rename_stack not only useless but also hazardous, since they would point
  to garbage!

- We have carefully designed to acquire oplog locks either singly or pair-wise
  (in a well-defined order), but never in an unbounded chain, making it quite
  straight-forward to get locking right!