bin: Bare-minimum quick-n-dirty test for rename and fsync
Just to verify that rename + fsync doesn't crash the kernel! ;-)

More seriously though, the new dependency handling code for rename
is kinda sophisticated. This basic test case helps us instrument the
kernel and verify that the new dependency handling code takes the expected
path and goes through the correct set of state changes, as we invoke
fsync on the source and destination directories.

[ Basic code structure borrowed from bin/fsynctest.cc ]