scalefs: Fix crash in dir_init() by supplying inum (rather than mnum) to iget()
iget() takes an inum as the second parameter, not an mnum. As a result of the
bug, dirlookup() would end up crashing in dir_init(), because the value
returned by iget() would have been invalid.

This bug manifested itself as "panic: dir_init not DIR" upon trying to
fsync the root directory after creating a file in it.