bin/renamefsync: Fsync the parent directory to flush all the paths
Under the new semantics for fsync, if the parent directory is not fsync()ed,
the paths to the newly created files or directories underneath it will not
be made persistent on the disk (and hence they won't be reachable by path
names on the disk after a reboot).

So fsync the parent directory in the test. [ We had skipped it because
of a bug that used to crash the kernel earlier; but the previous commit
fixed it. ]