scalefs: Teach process_ops_from_oplog() to process a specified number of operations
While processing a 'link' operation, we will have to process the 'create'
operation of the mnode which is being linked, as a dependency (so that its
corresponding inode will be created and initialized).

To achieve that within the current framework, we note that the 'create' operation
is always the first one logged for a given mnode; hence we modify the function
process_ops_from_oplog() to accept a count as an argument, and then process only
the first 'count' number of operations from the mfs_log. As expected, in a
subsequent patch, we will set count to 1 when trying to process 'create' operation
as a dependency for 'link'.