scalefs: Add an explicit dependency between 'link' and 'create' operations
The MFS 'link' operation needs to know the inode number of the mnode being
linked, and since the 'create' operation allocates and initializes inodes
for mnodes, it follows that 'link' depends on the 'create' operation.

However, the existing code used to bypass the need to explicitly flush out
the 'create' before the 'link' operation, by using a bit of trickery while
applying the 'link' operation (basically, by allocating an inode on-demand
from within the 'link' operation itself, and letting 'create' continue from
that point when it is processed later).

But it would be conceptually much more easier to understand if we have a
clear-cut dependency between 'link' and 'create' in the implementation. So
add an explicit dependency between the two, such that the mnode's 'create'
is processed before that mnode is linked anywhere.

In the current framework, this involves instructing process_ops_from_oplog()
to process only the first operation in the oplog (which is likely to be the
create operation), by setting count to 1. (We make this a special case in the
code, so that we don't process the first operation if it isn't a 'create';
this helps us keep the amount of disk-writes due to fsync at a minimum).