scalefs: Fix crash after rename due to use-after-free of transaction
We process the 'rename' operation by bundling both the link and the unlink
sub-operations into a single atomic transaction. However, the transaction used
to get added to the in-memory journal's list of transactions twice, i.e., after
applying each of the sub-operations. As a result, flush_journal_locked() used
to crash shortly after processing a rename, because it used to flush out the
rename transaction and delete it, and immediately try to flush out the same
transaction once again.

Fix this by adding the transaction to the journal's list of transactions only
after both the sub-operations of rename have been fully applied.