scalefs: Update the link to the parent atomically when renaming a directory
When a directory is moved from one parent directory to another, its ".." link
should also be updated to reflect the new parent. However, this change should
go along with the rename operation itself, in one single atomic transaction.

Luckily, the dependency handling for 'rename_link' and 'rename_unlink'
operations will ensure that the 'create' operation (which allocates inodes for
mnodes) would have been already processed for the directory being moved, by
the time we try to apply the rename operation. Thus, finding the inode of the
mdir and manipulating its ".." link is very straight-forward. We squeeze this
step along with rename_link, so that it gets applied within the same higher-
level rename transaction.