scalefs: Document the challenges involving fsync and directory-renames
Note down the scenarios involving directory renames and fsync, which can
potentially lead to inconsistencies in the filesystem, such as orphaned
loops on the disk.

Also propose a solution to this class of problems and write down a
semi-formal proof of its correctness.

This algorithm has not been implemented in our fsync yet.