scalefs: Robustify the directory-rename dependency tracking algorithm for fsync
The scenarios involving directory renames and fsync can be a little trickier
than what I had imagined earlier. So enhance our algorithm design so that it
can withstand those situations as well.

This patch adds the scenario description as well as the augmented solution
and its correctness proof to our design document.