scalefs, doc: Note down a complex scenario involving directory renames and fsync
Write down an example scenario where trying to naively avoid causing loops in
one part of the sub-tree might end up inadvertently causing loops in a different
sub-tree in the filesystem hierarchy on the disk.

This also serves as a testing ground for the new enhancement (recursion) that
we added to our fsync algorithm design.

Based on this example, make some minor changes to our algorithm to precisely
define its behavior on such an input. (The changes are not a whole lot; note
that the diff involving point *** has only whitespace changes).