scalefs, doc: Show that we cannot rely on timestamps to order rename dependencies
Given a directory to be fsynced, our algorithm might go up the filesystem tree
in memory from that directory and note down the set of directory-renames to be
flushed, as dependencies (to avoid creating loops on the disk). But in what
order should we flush those rename operations to disk? One might be tempted to
think that flushing them out in timestamp order will always be safe, but it turns
out that that isn't the case!

So, add a concrete example to show why it could be disastrous to rely solely on
timestamp order while flushing out dependent directory-rename operations.