scalefs, doc: Show an example where ancestor-descendant relationships are not rigid
One approach to avoid causing loops on the disk due to directory-renames and
fsync, is to keep track of which directories are the ancestors of a given
directory and then try to catch cases which attempt to trick fsync into moving
an ancestor into a descendant.

However, this approach won't really work that well in the long run, because
we can have scenarios where a directory A is an ancestor of directory B;
and after a few (valid) directory renames in memory, A is a descendant of B; and
then, after a few more (valid) transformations to the directory hierarchy, A is
once again an ancestor of B. This makes it quite complicated to keep track of
who is the ancestor and who is the descendant over a long series of renames in
memory.

This might not sound all that surprising, but the example shown below actually
brings out certain subtle pitfalls to avoid while designing an fsync algorithm
that aims to prevent loops on the disk.