scalefs/doc: Update the fsync loop-avoidance algorithm to handle design constraints
The previous version of the fsync loop-avoidance algorithm needed help from
MemFS at the time of fsync, to determine the ancestors of the destination
directory. This implied a tight coupling between the MemFS and the Oplog/Fsync
layers, which is undesirable. We had consciously kept them well-separated thus
far, with a controlled flow in one direction (MemFS -> Oplog); and we don't
want to introduce a backward dependency just to handle directory renames.

So we improved our algorithm to log additional state in the Oplog at the time
of performing the rename operation in MemFS, so that fsync (invoked eventually)
can proceed completely independently, with no help from MemFS. So update the
writeup with this new algorithm and its corresponding new proof of correctness.
An added bonus is that we now describe the algorithm in pseudocode, since it
allows us to describe it (as well as all its intricacies) precisely, which is
immensely helpful when trying to prove its correctness properties.