scalefs, rename: Remove non-POSIX-compliant behavior in the rename system call
POSIX specifies that if the source and the destination paths point to the same
inode, then the rename() system call must return success with no further action.
So fix our implementation of rename, which tries to delete the source path,
leaving the destination path as it is.

However, POSIX does allow 'mv' to delete the source path in this case (although
it is not mandated). We will follow this approach, since Linux does the same
thing as well.