scalefs, rename: Abort if the destination directory is killed by unlink
If the destination directory is empty, then a concurrent rmdir (unlink) can
kill its hash table right when we are about to add the new directory entry
to it as part of the rename operation. So verify that the destination
directory's hash table is still alive, with its lock held, before proceeding
any further with the rename. Since the kill operation (from rmdir/unlink)
also holds this lock, this check is not racy.

(Note that this fix is necessary even for file renames.)