sysfile, rename: Prevent forming loops in the filesystem tree within MemFS
Directory renames can introduce cycles (also called "orphaned loops") in the
filesystem hierarchy if we are not being careful - specifically, if we move
an ancestor into a descendant. So perform loop-avoidance in the MemFS layer
by verifying that the rename won't cause cycles in the filesystem tree.

Concurrent directory renames can make loop-avoidance very complicated to
reason about (not only in the MemFS layer, but also in the DiskFS layer later
on). So to keep things simple, have a global filesystem-wide lock to allow
only 1 directory rename to proceed at a time. However, note that this still
allows directory renames to run in parallel with (any number of) file renames.