chainhash, rename: Augment replace_from() to handle directory renames
When performing directory renames, we need to update 3 entities atomically
(as opposed to two when doing file renames):
  - the source directory (remove the old directory entry)
  - the destination directory (add the new directory entry)
  - the subdirectory being moved (update the ".." link to point to the new
    parent)
  - [ Actually, we should also deal with the directory that was originally
      in the destination directory and was replaced by the rename; but we'll
      postpone that for later. ]

So enhance replace_from() to take additional arguments representing the
changes to be made to the subdirectory being moved. When used in practice,
ksubdir will be ".." and vsubdir will be the mnode number of the destination
directory.

Now that we have 3 directories to update atomically, acquire the corresponding
locks for their hash table buckets in the increasing order of their addresses,
and update their directory entries with those locks held.