sysfile, rename: Relax the constraints on the file types used with rename
POSIX specifies that directories can be renamed to directories, and
non-directories can be renamed to non-directories. However, our current
check takes this a bit too far and allows only files to be renamed to files
in the latter category. For example, this would prevent sockets from being
renamed to files, which is actually legal.

So fix this by relaxing the constraints on the file types as specified by the
spec. Thanks to Nickolai for catching this.