scalefs: Introduce rename barriers to perform fsync loop-avoidance on the disk
We introduce "rename-barriers" as a mechanism to help us identify and flush
the set of dependent directory renames during an fsync, in order to prevent
forming orphaned loops in the filesystem hierarchy on the disk, when directory
renames are involved.

The idea behind our fsync loop-avoidance algorithm is outlined below:

Suppose a directory rename operation succeeded in memory (MemFS) (in
particular, it did not cause any loops within MemFS). Now, as part of a
subsequent fsync, we would like to flush this rename operation to disk,
without causing any loops in the disk filesystem.

We know that to avoid forming loops, we have to prevent moving an ancestor
directory into one of its descendant directories. So if we could somehow
arrange that, just before flushing this rename, the ancestors of the
destination directory on the disk are exactly the same as the ancestors of
the destination directory in MemFS at the time of the rename, then flushing
this rename wouldn't cause a loop on the disk (since it didn't cause a loop
in MemFS, with the same set of ancestors).

But it is not always possible to achieve this, because intervening renames
and fsync()s can alter the filesystem state on the disk. However, it turns
out that if we simply ensure that the state of the destination's ancestors
in memory (at the time of the rename) is reflected on the disk just before
flushing the rename operation itself, then we still end up preventing loops
on the disk.

Rename-barriers help us achieve just that, by capturing a distributed snapshot
of the state of the destination's ancestors in memory at the time of the
rename, which can be flushed to the disk during a sync/fsync later on.
Rename-barriers are flushed recursively starting from the root directory
and going down the tree towards the destination directory. The directory rename
operation itself is flushed only after flushing all its associated rename
barriers (which are identified by their common timestamps). This downward flow
also allows us to flush the rename barriers corresponding to a directory rename
operation in many parts; they don't have to be all flushed out in one go, and
this flexibility turns out to be particularly useful when executing other
fsync()s on the ancestor directories.