scalefs: Replace 'inum_to_mnode' with an 'inum_to_mnum' hash-table
The inum_to_mnode hash-table caches sref<mnode> objects, which has subtle
implications on the lifetimes of the mnodes and hence complicates reference
count management. For example, it introduces odd dependencies like this one:
even after the mfile or mdir is unlinked from MFS, its onzero() method won't
fire until its corresponding entry is deleted from this hash-table. Further,
it would be actually nice to have that entry around until fsync() gets invoked
later on, but because of this unwanted coupling with the mnode's refcount, we
have to free up its entry sort of prematurely.

All we need is a quick way to get to the mnode, given an inode number. So it
would be easier to just have a hash-table that maps inode numbers to mnode
numbers, and then we could just use the mnode_cache to lookup the corresponding
mnode. And we can keep the inum-to-mnum mapping as long as we want, without any
side-effects on the mnode's lifetime. Granted, this will now require two lookups
instead of one to get to the mnode from the inode number, but these lookups are
quite infrequent and are not in performance critical paths, so the slight
slowdown won't matter.