fs: Deferred deletion/reclamation of inodes with help from the superblock
When the last link to a file or directory goes away, if userspace still has open
file descriptors referring to it, then its inode cannot be deleted from the disk
during an fsync until all those open file descriptors get closed. In such cases,
we would like to note down the inode number in the filesystem metadata and then
reclaim such inodes upon reboot (deferred delete/reclaim).

In DiskFS, we place the superblock in its own separate disk block, but it
occupies a very small fraction of the block size (12/4096 bytes). So squeeze in
this list of inodes to be reclaimed within the superblock itself. This limits
the number of inodes we can mark for deferred deletion to a fixed constant, but
this scheme is good enough for now.