mnode, unlink: Use the mnode's refcount to decide whether the mnode has open FDs
Between the last unlink of the file or directory and the call to fsync(), a
refcache epoch might not have elapsed yet, which means that even if there are
no open file descriptors to that mnode, it can still be in the mnode_cache
waiting to be evicted. So look at the reference count of the mnode instead, to
get a more accurate picture of whether it is safe to delete its on-disk inode
or not.