fs: Don't log zeroed out block in the transaction during block allocation
We zero blocks during allocation, but we don't have to log them in the
transaction. An allocated block will be written to with useful data before being
logged in the transaction anyway, so it is wasteful to log the zeroed out block
as well (which will be thrown out by the deduplication code in any case).