fs: Return unlocked inode from ialloc()
There is no reason why we should return a locked inode from ialloc(); the caller
immediately unlocks it anyway, and moreover the write-lock is supposed to protect
writes to the inode (and the caller doesn't immediately write to it). Every other
function in this file returns unlocked inodes as well, as mentioned in the comments
outlining the locking design. And while at it, do some coding style fixes across
the file.