fs: Make ialloc() return a locked inode again
Although in our current implementation, ialloc() doesn't need to return a
locked inode because we process create()s strictly before any other operation,
it seems that things are easier to reason about if our correctness didn't
depend subtly on the order in which fsync() processes the operations.
So make ialloc() return a locked inode, so that we can be sure that a freshly
created inode will be added to the transaction in exactly the state that we
want, without any other writes interleaving with it.