fs: Don't waste time zero-filling the journal upon every transaction commit
In the current design, after a (group) transaction commits, we "clear" the
journal by zeroing it out before starting to log the next transaction in the
journal, so that we don't have stale transactions left over in it if we happen
to crash at that point.

However, this is quite inefficient (since the journal is about 4 MB and is made
up of more than a 1000 blocks) and unnecessary too. Each transaction already
embeds timestamp information in all its headers and commit blocks, so we can
simply use that info to avoid applying stale, corrupted or partial transactions
from the journal upon crash and reboot. Also, the process_journal() code which
is run at boot aborts if it finds the first header block to be a bunch of zeros.
So just add a zero header at the beginning to reset the journal. As a pleasant
side effect, we get to throw away the zero_fill() function altogether.