fs, drop_bufcache(): Abort if indirect and doubly-indirect blocks are not cached
If the metadata blocks (indirect and doubly-indirect blocks) of the inode are
themselves not in the buffer-cache, none of the data blocks they point to will
be in the bufcache either. So add checks to avoid pointless disk reads.

While at it, move the inode locking responsibility to the caller, just like
every other function in the file.