fs: Rewrite itrunc() in a more structured manner for readability
The current implementation of itrunc() is rather dense and riddled with lots of
special-case checks that makes it quite hard to decipher. Rewrite it in a more
structured way and also outline the locking protocol. Don't bother dealing with
ip->iaddrs, because we will get rid of it soon anyway.