fs: Make the inode's addrs[] array non-atomic and rewrite bmap() accordingly
We know that there can only be one writei() at a time on a given inode, so there
is no concurrency involved, and hence just regular non-atomic variables will do
just fine to hold the addresses of the inode's data blocks. Rewrite bmap() to
accommodate this change and throw away all its complexity: the loops, retries
and all the atomic compare-exchange operations.