fs: Remove double-checked locking from bmap() for cache-efficiency
Since bufcache uses seqlocks, calling the read() method on a bufcache object
will actually copy the entire block! (Note that BSIZE = 4096 bytes). For the
purpose of bmap(), we know that there is no concurrency, so it is more efficient
to directly acquire the uncontended write-lock on the bufcache block rather than
pollute the cache by reading 4K bytes every time.