scalefs: Don't use asynchronous disk I/O during early boot
The scheduler panics ("EMBRYO -> 1") if we try to use asynchronous disk I/O
during early boot (before calling idleloop()), because the process won't be
in a state that can be put to sleep to wait for the async I/O to complete.
(Unfortunately, this also implies that we can't use ideflush() in this path
at the moment, because it waits on condition variables too).

So rewrite the early boot code that processes the journal to use synchronous
disk I/O instead.