oplog: Reset obj_ pointer in synchronize_upto_tsc() to fix memory corruption
When free_metadata_log() calls delete on the mfs_logical_log pointer, at some
later point in time, the kernel crashes in the rmap's logging path: add_pte() ->
add_mapping() -> get_logger(). Within get_logger(), it crashes specifically at
this line:

  auto sync_guard = cur_obj->sync_lock_.try_guard();

Here, rmap has discovered that its hash-way is not free in the cache_, and is
trying to evict the existing object in order to take its place. To evict the
existing object, one has to acquire the object's sync_lock_ and flush the
logger.

But the system crashes because it turns out that cur_obj points to garbage!
Interestingly, way->obj_ (i.e., cur_obj in this case), is always updated with
the way's lock held, and get_logger() holds this same lock when reading way->obj_.
Further, when an mfs_logical_log object is deleted, its grandparent's destructor
(i.e, ~tsc_logged_object()) calls clear_loggers(), which acquires the way->lock
before setting way->obj_ to nullptr.

So there is no way that cur_obj could be pointing to garbage, right? Wrong!!
Here's why: clear_loggers() loops through all the cpus in the object's cpus_
bitmap and resets the way->obj_ pointer in each of them. A bit in the cpus_
bitmap is set whenever the object's operation is logged on that cpu (via
get_logger()), and is reset when the object is evicted from that cpu's cache_,
or if it is removed via a call to synchronize(). And when resetting the cpu bit
in both these scenarios, the way->obj_ pointer is also reset.

rmap uses synchronize(), whereas mfs_logical_log uses a special implementation
called synchronize_upto_tsc(), which resets the cpu in the bitmap, but fails
to reset the way's obj_ pointer! This means that this particular obj_ pointer
will be left around on this cpu's cache_, but this cpu itself will be invisible
to clear_loggers() because its bit has been reset. That's why rmap has trouble
evicting the object from the cache_'s way, because the stale obj_ pointer was
never reset on that cpu.

This also explains why rmap has trouble doing eviction whenever mfs_logical_log
is deleted, but not the other way around; because rmap uses synchronize(),
whereas mfs_logical_log uses the buggy synchronize_upto_tsc().

Fix this by resetting the obj_ pointer in synchronize_upto_tsc(). This was yet
another crazy bug that took several days to debug!