oplog: Document why sync_lock_ can afford to be a sleeplock
obj->sync_lock_ is a sleeplock while way->lock_ is a spinlock. Nesting the
spinlock acquisition inside the sleeplock acquisition is legal, whereas the
other way around would lead to what's known as the "sleeping inside atomic
section" bug.

The reason why this arrangement works in oplog is because, in synchronize()
and friends, we always nest the spinlock inside the sleeplock. However, in
get_logger(), we actually try to acquire them in the opposite order (which
would have panicked the kernel if we actually went through with it). But
in an effort to do deadlock avoidance, we end up acquiring the sleeplock as
if it is a spinlock (i.e., by retrying on contention, instead of sleeping).
So, all in all, this scheme is legal and hence works just fine. But there is
quite a bit of subtlety here and warrants some documentation.