sysfile: Cleanup numerous checks for root_fs
There are numerous checks in the syscall code to log metadata operations only
if we are operating on root_fs. Since that's the only case there is (at least as
of now), let us simplify the code by getting rid of those checks and adding
simple assert() statements instead. This cleanup will also help us with some
upcoming changes.