sysfile: Fix the placement of metadata_op_start() calls during directory rename
For any operation, we must maintain the invariant that metadata_op_start() is
called before performing the operation in MemFS, and metadata_op_end() is called
after. This ensures that the linearization point of the operation lies strictly
between the start() and the end() pairs. This in turn helps oplog's synchronize
family of functions to return the correct set of operations for fsync() at a
given timestamp, waiting for in-flight operations to get logged, if needed.

In the case of directory renames, the logging of rename barriers breaks this
invariant - the _start() calls are invoked on the destination's parent
directories *after* actually performing the rename! Fix this by invoking them
beforehand. This won't lead to any complications, because the directory hierarchy
is guaranteed not to change midway, since we hold a global lock during directory
renames in MemFS.