scalefs: Fix the logging of MFS operations in the face of thread migration
The MFS operation logging code is riddled with assumptions that the thread will
never get migrated to another cpu in between, nor will another thread take over
its place on the current cpu, which might not be true if the scheduler is to
have the flexibility to load-balance threads!

For starters, myid() is invoked separately for metadata_op_start() and
metadata_op_end(), which means that if the thread gets migrated between these
2 calls, then the operation's start tsc will be recorded on one cpu and its
corresponding end tsc will be recorded on another cpu! That is sure to wreak havoc
on the oplog's synchronize() code in the fsync() path, since the start and end
timestamps on a given cpu won't match (so synchronize_upto_tsc() can potentially
wait forever)!

Further, a write-lock is held while updating the start and end timestamps on
each cpu, but the write-lock is dropped in-between them! Which means that due to
load-balancing, we can have the following scenario:

Consider parallel file creates in a single directory:

              CPU 0                                            CPU1

// Preparing to create a new file                // Preparing to create a new file
// in dir A (dir mnum = 100).                    // in dir A (dir mnum = 100)

metadata_op_start(mnum100, cpu0, tsc1)                    |
                                                          |
Add file1 to dir A in MemFS (tsc1)                        |
                                                          |
                                                    <-----+ (migrated to CPU0)

// CPU1's thread got migrated to CPU0 and
// starts running, obliterating the earlier
// start timestamp!

metadata_op_start(mnum100, cpu0, tsc2)

Add file2 to dir A in MemFS (tsc2)

Log addition of file2 to mnum 100's oplog

metadata_op_end(mnum100, cpu0, tsc2)

...
...
...

As we can see above, the thread that migrated from CPU1 to CPU0 totally hijacked
the start and end timestamps from the thread that was originally running on CPU0.
Further, CPU0's thread had actually finished performing the MFS operation (linking
file1) on mnum 100 before the migration, which means that its linearization point
is actually ordered before that of file2's link into mnum 100. But file2's link
got logged in mnum 100's oplog along with its own end timestamp, without logging
file1's creation first. So when we call synchronize_upto_tsc() on mnum 100's oplog
as part of an fsync() on dir A, it can potentially return an incorrect set of
logged operations, by missing operations with earlier timestamps (such as file1's
create in this example), thus breaking our invariants!

Fix this by acquiring a per-oplog per-cpu write-lock before calling
metadata_op_start() and releasing it only after logging the MFS operation and
calling metadata_op_end(). That way, even if threads get migrated due to
load-balancing, our oplog and fsync invariants are maintained correctly.

To achieve this, we will also have to add a CPU argument to get_logger() so that
it logs the operation on the correct CPU, which was determined at the time of
calling metadata_op_start().

We build the necessary infrastructure in this patch and convert the callsites
in the next patch for clarity.