scalefs: Convert call-sites to fix MFS operation logging during thread migration
Now that we have the adequate infrastructure in place, convert all the places
in the syscall code where we log operations to the oplog.

The new scheme is:

1. Fix the cpu number by caching the return value of myid() locally.
2. Acquire the per-oplog per-cpu lock (which protects the start and end
   timestamps) by calling metadata_op_lockguard()

Then, perform the following operations by supplying the same cpu number as
argument while keeping the above lock alive:

 * metadata_op_start()
 * Perform the operation in MemFS
 * add_to_metadata_log()
 * metadata_op_end()

This gets a bit tricky when a single MFS operation involves logging in different
oplogs such as create() and particularly rename(), so we take care to acquire
all the relevant locks in a consistent order (ordered by increasing mnode
numbers) to avoid deadlocks.

This scheme does not hurt scalability since these locks are not shared; they are
per-oplog and per-cpu.