sysfile, rename: Remove outdated comment about ordering calls to metadata_op_start()
We call metadata_op_start() on all the relevant directories *before* performing
the rename in MemFS. So it automatically follows that if either of the
rename-link or rename-unlink sub-operations are found in the oplog after a call
to synchronize_upto_tsc(), the other one will also be found, because the start
timestamps are recorded before performing the actual operation in MemFS and hence
synchronize_upto_tsc() will wait for the operation to get logged.

Furthermore, this invariant (that synchronize_upto_tsc() will find all the parts
of a given rename operation) is even more crucial for cross-directory directory
renames because, apart from the rename-link and rename-unlink operation, we also
log a whole bunch of rename-barriers in several different directories (ancestors
of the destination) in order to avoid loops on fsync. So it is nice to have one
simple scheme that works beautifully for everything.

Thus, the order in which we call metadata_op_start() on the source and
destination parent directories doesn't matter, as long as all the calls to
metadata_op_start() happen before we actually perform the rename in MemFS.